made the controller logic clear for the all the endpoints
